Improved undefined behavior fix

+++ src/EratMedium.hpp
@@ -1,7 +1,7 @@
 ///
 /// @file  EratMedium.hpp
 ///
-/// Copyright (C) 2025 Kim Walisch, <kim.walisch@gmail.com>
+/// Copyright (C) 2026 Kim Walisch, <kim.walisch@gmail.com>
 ///
 /// This file is distributed under the BSD License. See the COPYING
 /// file in the top level directory.
@@ -32,7 +32,7 @@ class EratMedium : public Wheel30_t
 public:
   void init(uint64_t, uint64_t, MemoryPool&);
   bool hasSievingPrimes() const { return !buckets_.empty(); }
-  NOINLINE void crossOff(Vector<uint8_t>& sieve);
+  NOINLINE void crossOff(Vector<uint64_t>& sieve);
 private:
   uint64_t maxPrime_ = 0;
   MemoryPool* memoryPool_ = nullptr;
